refactor: code-split into transform menu with ctx
